moved visualisations from core to mpl
