+++ pymove/core/pandas.py
@@ -12,8 +12,6 @@ from typing import (
 )
 
 import numpy as np
-from matplotlib import pyplot as plt
-from matplotlib.pyplot import figure
 from pandas import DataFrame, DateOffset, Series, Timedelta
 
 from pymove.core.dataframe import MoveDataFrame
@@ -1383,223 +1381,6 @@ class PandasMoveDataFrame(DataFrame):
         self.last_operation = end_operation(operation)
 
         return bbox_
-
-    def plot_all_features(
-        self,
-        dtype: Optional[Callable] = np.float64,
-        figsize: Optional[Tuple[float, float]] = (21, 15),
-        return_fig: Optional[bool] = True,
-        save_fig: Optional[bool] = False,
-        name: Optional[Text] = 'features.png',
-    ) -> Optional[figure]:
-        """
-        Generate a visualization for each columns that type is equal dtype.
-
-        Parameters
-        ----------
-        dtype : callable, optional
-            Represents column type, by default np.float64
-        figsize : tuple(float, float), optional
-            Represents dimensions of figure, by default (21, 15)
-        return_fig : bool, optional
-            Represents whether or not to return the generated picture, by default True
-        save_fig : bool, optional
-            Represents whether or not to save the generated picture, by default False
-        name : str, optional
-            Represents name of a file, by default 'features.png'
-
-        Returns
-        -------
-        figure
-            The generated picture or None
-
-        Raises
-        ------
-        AttributeError
-            If there are no columns with the specified type
-
-        """
-
-        operation = begin_operation('plot_all_features')
-
-        try:
-            col_dtype = self.select_dtypes(include=[dtype]).columns
-            tam = col_dtype.size
-            if not tam:
-                raise AttributeError('No columns with dtype %s.' % dtype)
-
-            fig, ax = plt.subplots(tam, 1, figsize=figsize)
-            ax_count = 0
-            for col in col_dtype:
-                ax[ax_count].set_title(col)
-                self[col].plot(subplots=True, ax=ax[ax_count])
-                ax_count += 1
-
-            if save_fig:
-                plt.savefig(fname=name, fig=fig)
-
-            self.last_operation = end_operation(operation)
-
-            if return_fig:
-                return fig
-        except Exception as e:
-            self.last_operation = end_operation(operation)
-            raise e
-
-    def plot_trajs(
-        self,
-        markers: Optional[Text] = 'o',
-        markersize: Optional[float] = 12,
-        figsize: Optional[Tuple[float, float]] = (10, 10),
-        return_fig: Optional[bool] = True,
-        save_fig: Optional[bool] = False,
-        name: Optional[Text] = 'trajectories.png',
-    ) -> Optional[figure]:
-        """
-        Generate a visualization that show trajectories.
-
-        Parameters
-        ----------
-        markers : str, optional
-             Represents visualization type marker, by default 'o'
-        markersize : float, optional
-            Represents visualization size marker, by default 12
-        figsize : tuple(float, float), optional
-             Represents dimensions of figure, by default (10, 10)
-        return_fig : bool, optional
-            Represents whether or not to return the generated picture, by default True
-        save_fig : bool, optional
-            Represents whether or not to save the generated picture, by default False
-        name : str, optional
-            Represents name of a file, by default 'trajectories.png'
-
-        Returns
-        -------
-        figure
-            The generated picture or None
-        """
-
-        operation = begin_operation('plot_trajs')
-
-        fig = plt.figure(figsize=figsize)
-
-        ids = self['id'].unique()
-        for id_ in ids:
-            self_id = self[self['id'] == id_]
-            plt.plot(
-                self_id[LONGITUDE],
-                self_id[LATITUDE],
-                markers,
-                markersize=markersize,
-            )
-
-        if save_fig:
-            plt.savefig(fname=name, fig=fig)
-
-        self.last_operation = end_operation(operation)
-
-        if return_fig:
-            return fig
-
-    def plot_traj_id(
-        self,
-        tid: Union[int, Text],
-        label: Optional[Text] = TID,
-        feature: Optional[Text] = None,
-        value: Optional[Any] = None,
-        markersize: Optional[float] = 20,
-        figsize: Optional[Tuple[float, float]] = (10, 10),
-        return_fig: Optional[bool] = True,
-        save_fig: Optional[bool] = False,
-        name: Optional[Text] = None,
-    ) -> Optional[figure]:
-        """
-        Generate a visualization that shows a trajectory with the specified tid.
-
-        Parameters
-        ----------
-        tid :  int, str
-            Represents the trajectory tid
-        label : str, optional
-            Feature with trajectories tids, by default TID
-        feature : str, optional
-            Name of the feature to highlight on plot, by default None
-        value : any, optional
-            Value of the feature to be highlighted as green marker, by default None
-        markersize : float, optional
-            Represents visualization size marker, by default 20
-        figsize : tuple(float, float), optional
-            Represents dimensions of figure, by default (10, 10)
-        return_fig : bool, optional
-            Represents whether or not to return the generated picture, by default True
-        save_fig : bool, optional
-            Represents whether or not to save the generated picture, by default False
-        name : str, optional
-            Represents name of a file, by default None
-
-        Returns
-        -------
-        PandasMoveDataFrame', figure
-            Trajectory with the specified tid.
-            The generated picture.
-
-        Raises
-        ------
-        KeyError
-            If the dataframe does not contains the TID feature
-        IndexError
-            If there is no trajectory with the tid passed
-
-        """
-
-        operation = begin_operation('plot_traj_id')
-
-        if label not in self:
-            self.last_operation = end_operation(operation)
-            raise KeyError('%s feature not in dataframe' % label)
-
-        df_ = self[self[label] == tid]
-
-        if not len(df_):
-            self.last_operation = end_operation(operation)
-            raise IndexError(f'No trajectory with tid {tid} in dataframe')
-
-        fig = plt.figure(figsize=figsize)
-
-        plt.plot(
-            df_.iloc[0][LONGITUDE], df_.iloc[0][LATITUDE], 'yo', markersize=markersize
-        )  # start point
-        plt.plot(
-            df_.iloc[-1][LONGITUDE], df_.iloc[-1][LATITUDE], 'yX', markersize=markersize
-        )  # end point
-
-        if (not feature) or (not value) or (feature not in df_):
-            plt.plot(df_[LONGITUDE], df_[LATITUDE])
-            plt.plot(
-                df_.loc[:, LONGITUDE], df_.loc[:, LATITUDE],
-                'r.', markersize=markersize / 2
-            )
-        else:
-            filter_ = df_[feature] == value
-            df_nodes = df_.loc[filter_]
-            df_points = df_.loc[~filter_]
-            plt.plot(df_[LONGITUDE], df_[LATITUDE], linewidth=3)
-            plt.plot(
-                df_nodes[LONGITUDE], df_nodes[LATITUDE], 'gs', markersize=markersize / 2
-            )
-            plt.plot(
-                df_points[LONGITUDE], df_points[LATITUDE], 'r.', markersize=markersize / 2
-            )
-
-        if save_fig:
-            if not name:
-                name = 'trajectory_%s.png' % tid
-            plt.savefig(fname=name, fig=fig)
-
-        self.last_operation = end_operation(operation)
-
-        if return_fig:
-            return fig
 
     def show_trajectories_info(self):
         """
